server/requirements.txt
absl-py==0.11.0
appdirs==1.4.4
astunparse==1.6.3
attrs==19.3.0
Automat==0.8.0
black==20.8b1
blinker==1.4
cachetools==4.2.1
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
cloud-init==20.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
Flask==1.1.2
Flask-Cors==3.0.10
flatbuffers==1.12
gast==0.3.3
google-auth==1.27.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==1.0.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
Markdown==3.3.3
MarkupSafe==1.1.0
more-itertools==4.2.0
mypy-extensions==0.4.3
netifaces==0.10.4
nltk==3.5
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
pathspec==0.8.1
pexpect==4.6.0
pickle-mixin==1.0.2
protobuf==3.15.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.4
python-debian===0.1.36ubuntu1
python-dotenv==0.15.0
PyYAML==5.3.1
regex==2020.11.13
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.1
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.15.0
sos==4.0
ssh-import-id==5.10
systemd-python==234
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
toml==0.10.2
tqdm==4.57.0
tweepy==3.10.0
Twisted==18.9.0
typed-ast==1.4.2
typing-extensions==3.7.4.3
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==1.0.0
zope.interface==4.7.1
